Checkpoint work towards getting images working in glue-jupyter
