+++ glue_wwt/viewer/table_style_editor.py
