+++ glue_wwt/viewer/viewer_state.py
@@ -1,12 +1,8 @@
 from __future__ import absolute_import, division, print_function
 
-from distutils.version import LooseVersion
-
 import pywwt
 
 from astropy import units as u
-
-from glue.config import colormaps
 
 from glue.external.echo import (CallbackProperty, ListCallbackProperty,
                                 SelectionCallbackProperty, delay_callback,
@@ -15,8 +11,6 @@ from glue.external.echo import (CallbackProperty, ListCallbackProperty,
 from glue.core.data_combo_helper import ComponentIDComboHelper
 from glue.core.state_objects import StateAttributeLimitsHelper
 from glue.viewers.common.state import ViewerState, LayerState
-
-PYWWT_LT_06 = LooseVersion(pywwt.__version__) < '0.6'
 
 MODES_BODIES = ['Sun', 'Mercury', 'Venus', 'Earth', 'Moon', 'Mars',
                 'Jupiter', 'Callisto', 'Europa', 'Ganymede', 'Io', 'Saturn',
@@ -107,110 +101,3 @@ class WWTDataViewerState(ViewerState):
             return 1
         else:
             return 0
-
-
-class WWTLayerState(LayerState):
-    """
-    A state object for volume layers
-    """
-
-
-    layer = CallbackProperty()
-    color = CallbackProperty()
-    size = CallbackProperty()
-    alpha = CallbackProperty()
-
-    size_mode = SelectionCallbackProperty(default_index=0)
-    size = CallbackProperty()
-    size_att = SelectionCallbackProperty()
-    size_vmin = CallbackProperty()
-    size_vmax = CallbackProperty()
-    size_scaling = CallbackProperty(1)
-
-    color_mode = SelectionCallbackProperty(default_index=0)
-    cmap_att = SelectionCallbackProperty()
-    cmap_vmin = CallbackProperty()
-    cmap_vmax = CallbackProperty()
-    cmap = CallbackProperty()
-    cmap_mode = color_mode
-
-    size_limits_cache = CallbackProperty({})
-    cmap_limits_cache = CallbackProperty({})
-
-    def __init__(self, layer=None, **kwargs):
-
-        self._sync_markersize = None
-
-        super(WWTLayerState, self).__init__(layer=layer)
-
-        self._sync_color = keep_in_sync(self, 'color', self.layer.style, 'color')
-        self._sync_alpha = keep_in_sync(self, 'alpha', self.layer.style, 'alpha')
-        self._sync_size = keep_in_sync(self, 'size', self.layer.style, 'markersize')
-
-        self.color = self.layer.style.color
-        self.size = self.layer.style.markersize
-        self.alpha = self.layer.style.alpha
-
-        self.size_att_helper = ComponentIDComboHelper(self, 'size_att',
-                                                      numeric=True,
-                                                      categorical=False)
-        self.cmap_att_helper = ComponentIDComboHelper(self, 'cmap_att',
-                                                      numeric=True,
-                                                      categorical=False)
-
-        self.size_lim_helper = StateAttributeLimitsHelper(self, attribute='size_att',
-                                                          lower='size_vmin', upper='size_vmax',
-                                                          cache=self.size_limits_cache)
-
-        self.cmap_lim_helper = StateAttributeLimitsHelper(self, attribute='cmap_att',
-                                                          lower='cmap_vmin', upper='cmap_vmax',
-                                                          cache=self.cmap_limits_cache)
-
-        self.add_callback('layer', self._on_layer_change)
-        if layer is not None:
-            self._on_layer_change()
-
-        self.cmap = colormaps.members[0][1]
-
-        # Color and size encoding depending on attributes is only available
-        # in PyWWT 0.6 or later.
-        if PYWWT_LT_06:
-            modes = ['Fixed']
-        else:
-            modes = ['Fixed', 'Linear']
-
-        WWTLayerState.color_mode.set_choices(self, modes)
-        WWTLayerState.size_mode.set_choices(self, modes)
-
-        self.update_from_dict(kwargs)
-
-    def _on_layer_change(self, layer=None):
-
-        with delay_callback(self, 'cmap_vmin', 'cmap_vmax', 'size_vmin', 'size_vmax'):
-
-            if self.layer is None:
-                self.cmap_att_helper.set_multiple_data([])
-                self.size_att_helper.set_multiple_data([])
-            else:
-                self.cmap_att_helper.set_multiple_data([self.layer])
-                self.size_att_helper.set_multiple_data([self.layer])
-
-    def update_priority(self, name):
-        return 0 if name.endswith(('vmin', 'vmax')) else 1
-
-    def _layer_changed(self):
-
-        super(WWTLayerState, self)._layer_changed()
-
-        if self._sync_markersize is not None:
-            self._sync_markersize.stop_syncing()
-
-        if self.layer is not None:
-            self.size = self.layer.style.markersize
-            self._sync_markersize = keep_in_sync(self, 'size', self.layer.style, 'markersize')
-
-    def flip_size(self):
-        self.size_lim_helper.flip_limits()
-
-    def flip_cmap(self):
-        self.cmap_lim_helper.flip_limits()
